set_cpu_control_impl
